fix a warning on osx
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@6515 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ sched.c
@@ -316,7 +316,7 @@ void ast_sched_dump(const struct sched_context *con)
 			q->callback,
 			q->data,
 			delta.tv_sec,
-			delta.tv_usec);
+			(long int)delta.tv_usec);
 	}
 	ast_log(LOG_DEBUG, "=============================================================\n");
